requirements.txt
# local package
-e .

# external requirements
click==8.0.3
dill==0.3.4
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.21.2
pandas==1.3.3
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
scikit-learn==1.0
scipy==1.7.3
six==1.16.0
threadpoolctl==3.0.0
Werkzeug==2.0.2
requests==2.27.1
gunicorn==20.1.0